% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_greenmaps.R
\name{appdir}
\alias{appdir}
\title{Get GreenMaps Application Directory}
\usage{
appdir()
}
\value{
A character string representing the file path to the top directory
of the \emph{GreenMaps} Shiny app.
}
\description{
This function returns the path to the top directory
of the \emph{GreenMaps} Shiny application.
}
\examples{
appdir()
}
