% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_greenmaps.R
\name{run_GreenMaps}
\alias{run_GreenMaps}
\title{Run \emph{GreenMaps}}
\usage{
run_GreenMaps(launch.browser = TRUE, port = getOption("shiny.port"))
}
\arguments{
\item{launch.browser}{Whether or not to launch a new browser window.}

\item{port}{The port for the shiny server to listen on. Defaults to a
random available port.}
}
\description{
This function runs the \emph{GreenMaps} shiny application
}
\examples{
if(interactive()) {
run_GreenMaps()
}
}
\author{
Barnabas Daru <darunabas@gmail.com>
}
